overflow bug fix

+++ algorithmic-toolbox/week6/PlacingParentheses.java
@@ -5,22 +5,22 @@ import java.util.Scanner;
  * O(n^2) extra space
  */
 public class PlacingParentheses {
-    private static int[] getMaxMin(int[][] M, int[][] m, char[] operators, int i, int j) {
-	int max = Integer.MIN_VALUE, min = Integer.MAX_VALUE;
+    private static long[] getMaxMin(long[][] M, long[][] m, char[] operators, int i, int j) {
+	long max = Long.MIN_VALUE, min = Long.MAX_VALUE;
 	for (int k = i; k <= j - 1; k++) {
-	    int a1 = eval(M[i][k], M[k + 1][j], operators[k]);
-	    int a2 = eval(M[i][k], m[k + 1][j], operators[k]);
-	    int a3 = eval(m[i][k], M[k + 1][j], operators[k]);
-	    int a4 = eval(m[i][k], m[k + 1][j], operators[k]);
-	    int maxK = Math.max(Math.max(Math.max(a1, a2), a3), a4);
-	    int minK = Math.min(Math.min(Math.min(a1, a2), a3), a4);
+	    long a1 = eval(M[i][k], M[k + 1][j], operators[k]);
+	    long a2 = eval(M[i][k], m[k + 1][j], operators[k]);
+	    long a3 = eval(m[i][k], M[k + 1][j], operators[k]);
+	    long a4 = eval(m[i][k], m[k + 1][j], operators[k]);
+	    long maxK = Math.max(Math.max(Math.max(a1, a2), a3), a4);
+	    long minK = Math.min(Math.min(Math.min(a1, a2), a3), a4);
 	    if (maxK > max)
 		max = maxK;
 	    if (minK < min)
 		min = minK;
 	}
 
-	int maxmin[] = new int[] { max, min };
+	long maxmin[] = new long[] { max, min };
 	return maxmin;
     }
 
@@ -35,7 +35,7 @@ public class PlacingParentheses {
 		operators[i / 2] = exp.charAt(i);
 	}
 
-	int[][] M = new int[n][n], m = new int[n][n];
+	long[][] M = new long[n][n], m = new long[n][n];
 	for (int i = 0; i < n; i++) {
 	    M[i][i] = operands[i];
 	    m[i][i] = operands[i];
@@ -44,7 +44,7 @@ public class PlacingParentheses {
 	for (int s = 1; s < n; s++) {
 	    for (int i = 0; i < n - s; i++) {
 		int j = i + s;
-		int[] maxmin = getMaxMin(M, m, operators, i, j);
+		long[] maxmin = getMaxMin(M, m, operators, i, j);
 		M[i][j] = maxmin[0];
 		m[i][j] = maxmin[1];
 	    }
@@ -53,7 +53,7 @@ public class PlacingParentheses {
 	return M[0][n - 1];
     }
 
-    private static int eval(int a, int b, char op) {
+    private static long eval(long a, long b, char op) {
 	if (op == '+') {
 	    return a + b;
 	} else if (op == '-') {
